requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dlib-bin @ file:///D:/AI/py/Web/dlib_bin-19.24.2-cp39-cp39-win_amd64.whl#sha256=3ac61a810aa963510dc6dba550750dbb5224fcb02072e2663888732f4fae4619
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
fsspec==2025.3.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.134
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
